+++ packages/cinterop/src/darwin/kotlin/io/realm/internal/interop/RealmInterop.kt
@@ -19,6 +19,8 @@
 package io.realm.internal.interop
 
 import io.realm.internal.interop.Constants.ENCRYPTION_KEY_LENGTH
+import io.realm.internal.interop.RealmInterop.propertyInfo
+import io.realm.internal.interop.RealmInterop.safeKString
 import io.realm.internal.interop.sync.AuthProvider
 import io.realm.internal.interop.sync.CoreUserState
 import io.realm.internal.interop.sync.MetadataMode
@@ -29,11 +31,14 @@ import io.realm.mongodb.SyncException
 import kotlinx.cinterop.BooleanVar
 import kotlinx.cinterop.ByteVar
 import kotlinx.cinterop.ByteVarOf
+import kotlinx.cinterop.CArrayPointer
 import kotlinx.cinterop.COpaquePointer
 import kotlinx.cinterop.CPointed
 import kotlinx.cinterop.CPointer
 import kotlinx.cinterop.CPointerVar
 import kotlinx.cinterop.CValue
+import kotlinx.cinterop.CVariable
+import kotlinx.cinterop.LongVar
 import kotlinx.cinterop.MemScope
 import kotlinx.cinterop.StableRef
 import kotlinx.cinterop.UIntVar
@@ -153,6 +158,10 @@ fun realm_value_t.set(memScope: MemScope, value: Any?): realm_value_t {
         is String -> {
             type = realm_value_type.RLM_TYPE_STRING
             string.set(memScope, value)
+        }
+        is Boolean -> {
+            type = realm_value_type.RLM_TYPE_BOOL
+            boolean = value
         }
         is Byte, is Short, is Int, is Long -> {
             type = realm_value_type.RLM_TYPE_INT
@@ -616,6 +625,10 @@ actual object RealmInterop {
         return realm_wrapper.realm_object_is_valid(obj.cptr())
     }
 
+    actual fun realm_object_get_key(obj: NativePointer): Long {
+        return realm_wrapper.realm_object_get_key(obj.cptr())
+    }
+
     actual fun realm_object_resolve_in(obj: NativePointer, realm: NativePointer): NativePointer? {
         memScoped {
             val objectPointer = allocArray<CPointerVar<realm_object_t>>(1)
@@ -869,7 +882,7 @@ actual object RealmInterop {
         memScoped {
             val count = args.size
             val cArgs = allocArray<realm_value_t>(count)
-            args.mapIndexed { i, arg ->
+            arg.mapIndexed { i, arg ->
                 cArgs[i].apply {
                     set(memScope, arg)
                 }
@@ -1198,6 +1211,97 @@ actual object RealmInterop {
             ),
             managed = false
         )
+    }
+
+    actual fun realm_object_changes_get_modified_properties(change: NativePointer): List<PropertyKey> {
+        val propertyCount = realm_wrapper.realm_object_changes_get_num_modified_properties(change.cptr())
+
+        memScoped {
+            val propertyKeys = allocArray<LongVar>(propertyCount.toLong())
+            realm_wrapper.realm_object_changes_get_modified_properties(change.cptr(), propertyKeys, propertyCount)
+            return (0 until propertyCount.toInt()).map { PropertyKey(propertyKeys[it].toLong()) }
+        }
+    }
+
+    private inline fun <reified T : CVariable> MemScope.initArray(size: CArrayPointer<ULongVar>) = allocArray<T>(size[0].toInt())
+
+    actual fun <T, R> realm_collection_changes_get_indices(change: NativePointer, builder: ListChangeSetBuilder<T, R>) {
+        memScoped {
+            val insertionCount = allocArray<ULongVar>(1)
+            val deletionCount = allocArray<ULongVar>(1)
+            val modificationCount = allocArray<ULongVar>(1)
+            val movesCount = allocArray<ULongVar>(1)
+
+            realm_wrapper.realm_collection_changes_get_num_changes(change.cptr(), deletionCount, insertionCount, modificationCount, movesCount)
+
+            val deletionIndices = initArray<ULongVar>(deletionCount)
+            val insertionIndices = initArray<ULongVar>(insertionCount)
+            val modificationIndices = initArray<ULongVar>(modificationCount)
+            val modificationIndicesAfter = initArray<ULongVar>(modificationCount)
+            val moves = initArray<realm_wrapper.realm_collection_move_t>(movesCount)
+
+            realm_wrapper.realm_collection_changes_get_changes(
+                change.cptr(),
+                deletionIndices,
+                deletionCount[0],
+                insertionIndices,
+                insertionCount[0],
+                modificationIndices,
+                modificationCount[0],
+                modificationIndicesAfter,
+                modificationCount[0],
+                moves,
+                movesCount[0]
+            )
+
+            builder.initIndicesArray(builder::insertionIndices, insertionCount, insertionIndices)
+            builder.initIndicesArray(builder::deletionIndices, deletionCount, deletionIndices)
+            builder.initIndicesArray(builder::modificationIndices, modificationCount, modificationIndices)
+            builder.initIndicesArray(builder::modificationIndicesAfter, modificationCount, modificationIndicesAfter)
+            builder.movesCount = movesCount[0].toInt()
+        }
+    }
+
+    actual fun <T, R> realm_collection_changes_get_ranges(change: NativePointer, builder: ListChangeSetBuilder<T, R>) {
+        memScoped {
+            val insertRangesCount = allocArray<ULongVar>(1)
+            val deleteRangesCount = allocArray<ULongVar>(1)
+            val modificationRangesCount = allocArray<ULongVar>(1)
+            val movesCount = allocArray<ULongVar>(1)
+
+            realm_wrapper.realm_collection_changes_get_num_ranges(
+                change.cptr(),
+                deleteRangesCount,
+                insertRangesCount,
+                modificationRangesCount,
+                movesCount
+            )
+
+            val insertionRanges = initArray<realm_wrapper.realm_index_range_t>(insertRangesCount)
+            val modificationRanges = initArray<realm_wrapper.realm_index_range_t>(modificationRangesCount)
+            val modificationRangesAfter = initArray<realm_wrapper.realm_index_range_t>(modificationRangesCount)
+            val deletionRanges = initArray<realm_wrapper.realm_index_range_t>(deleteRangesCount)
+            val moves = initArray<realm_wrapper.realm_collection_move_t>(movesCount)
+
+            realm_wrapper.realm_collection_changes_get_ranges(
+                change.cptr(),
+                deletionRanges,
+                deleteRangesCount[0],
+                insertionRanges,
+                insertRangesCount[0],
+                modificationRanges,
+                modificationRangesCount[0],
+                modificationRangesAfter,
+                modificationRangesCount[0],
+                moves,
+                movesCount[0]
+            )
+
+            builder.initRangesArray(builder::deletionRanges, deleteRangesCount, deletionRanges)
+            builder.initRangesArray(builder::insertionRanges, insertRangesCount, insertionRanges)
+            builder.initRangesArray(builder::modificationRanges, modificationRangesCount, modificationRanges)
+            builder.initRangesArray(builder::modificationRangesAfter, modificationRangesCount, modificationRangesAfter)
+        }
     }
 
     // TODO sync config shouldn't be null