+++ packages/library-base/src/jvm/kotlin/io/realm/internal/platform/RealmObject.kt
@@ -30,4 +30,4 @@ internal actual fun <T : Any> realmObjectCompanionOrNull(clazz: KClass<T>): Real
 
 internal actual fun <T : RealmObject> realmObjectCompanionOrThrow(clazz: KClass<T>): RealmObjectCompanion =
     realmObjectCompanionOrNull(clazz)
-        ?: error("Couldn't find companion object of class ${clazz.simpleName}")
+        ?: error("Couldn't find companion object of class '${clazz.simpleName}'.\nA common cause for this is when the `io.realm.kotlin` is not applied to the Gradle module that contains the '${clazz.simpleName}' class.")