+++ test/base/src/androidTest/kotlin/io/realm/test/shared/notifications/RealmResultsNotificationsTests.kt
@@ -19,15 +19,22 @@ package io.realm.test.shared.notifications
 import co.touchlab.stately.concurrency.AtomicInt
 import io.realm.Realm
 import io.realm.RealmConfiguration
-import io.realm.RealmResults
 import io.realm.entities.Sample
+import io.realm.entities.list.RealmListContainer
+import io.realm.notifications.InitialResults
+import io.realm.notifications.ListChangeSet
+import io.realm.notifications.ResultsChange
+import io.realm.notifications.UpdatedResults
 import io.realm.query
 import io.realm.query.find
 import io.realm.test.NotificationTests
+import io.realm.test.assertIsChangeSet
 import io.realm.test.platform.PlatformUtils
+import io.realm.test.shared.OBJECT_VALUES
+import io.realm.test.shared.OBJECT_VALUES2
+import io.realm.test.shared.OBJECT_VALUES3
 import kotlinx.coroutines.async
 import kotlinx.coroutines.channels.Channel
-import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.flow.filterNot
 import kotlinx.coroutines.runBlocking
 import kotlin.test.AfterTest
@@ -35,6 +42,7 @@ import kotlin.test.BeforeTest
 import kotlin.test.Ignore
 import kotlin.test.Test
 import kotlin.test.assertEquals
+import kotlin.test.assertIs
 import kotlin.test.assertNotNull
 import kotlin.test.assertTrue
 import kotlin.test.fail
@@ -48,7 +56,7 @@ class RealmResultsNotificationsTests : NotificationTests {
     @BeforeTest
     fun setup() {
         tmpDir = PlatformUtils.createTempDir()
-        configuration = RealmConfiguration.Builder(schema = setOf(Sample::class))
+        configuration = RealmConfiguration.Builder(schema = setOf(Sample::class, RealmListContainer::class))
             .path(path = "$tmpDir/default.realm").build()
         realm = Realm.open(configuration)
     }
@@ -64,7 +72,7 @@ class RealmResultsNotificationsTests : NotificationTests {
     @Test
     override fun initialElement() {
         runBlocking {
-            val c = Channel<RealmResults<Sample>>(1)
+            val c = Channel<ResultsChange<Sample>>(1)
             val observer = async {
                 realm.query<Sample>()
                     .asFlow()
@@ -72,30 +80,189 @@ class RealmResultsNotificationsTests : NotificationTests {
                         c.trySend(it)
                     }
             }
-            val initialElement: RealmResults<Sample> = c.receive()
-            assertEquals(0, initialElement.size)
+
+            c.receive().let { resultsChange ->
+                assertIs<InitialResults<*>>(resultsChange)
+                assertTrue(resultsChange.list.isEmpty())
+            }
+
             observer.cancel()
             c.close()
         }
     }
 
     @Test
+    @Suppress("LongMethod")
     override fun asFlow() {
+        val dataset1 = OBJECT_VALUES
+        val dataset2 = OBJECT_VALUES2
+        val dataset3 = OBJECT_VALUES3
+
         runBlocking {
-            val c = Channel<Int>(capacity = 1)
+            val c = Channel<ResultsChange<*>>(capacity = 1)
             val observer = async {
-                realm.query<Sample>()
+                realm.query<RealmListContainer>()
+                    .sort("stringField")
                     .asFlow()
-                    .filterNot {
-                        it.isEmpty()
-                    }.collect {
-                        c.trySend(it.size)
+                    .collect {
+                        c.trySend(it)
                     }
             }
-            realm.write {
-                copyToRealm(Sample().apply { stringField = "Foo" })
+
+            // Assertion after empty list is emitted
+            c.receive().let { resultsChange ->
+                assertIs<InitialResults<*>>(resultsChange)
+
+                assertNotNull(resultsChange.list)
+                assertEquals(0, resultsChange.list.size)
             }
-            assertEquals(1, c.receive())
+
+            // Assert a single range is reported
+            //
+            // objectListField = [C, D, E, F]
+            realm.writeBlocking {
+                dataset2.forEach {
+                    copyToRealm(it)
+                }
+            }
+
+            c.receive().let { resultsChange ->
+                assertIs<UpdatedResults<*>>(resultsChange)
+
+                assertNotNull(resultsChange.list)
+                assertEquals(dataset2.size, resultsChange.list.size)
+
+                assertIsChangeSet(
+                    (resultsChange as UpdatedResults<*>),
+                    insertRanges = arrayOf(
+                        ListChangeSet.Range(0, 4)
+                    )
+                )
+            }
+
+            // Assert multiple ranges are reported
+            //
+            // objectListField = [<A, B>, C, D, E, F, <G, H>]
+            realm.writeBlocking {
+                (dataset1 + dataset3).forEach {
+                    copyToRealm(it)
+                }
+            }
+
+            c.receive().let { resultsChange ->
+                assertIs<UpdatedResults<*>>(resultsChange)
+
+                assertNotNull(resultsChange.list)
+                assertEquals(dataset1.size + dataset2.size + dataset3.size, resultsChange.list.size)
+
+                assertIsChangeSet(
+                    (resultsChange as UpdatedResults<*>),
+                    insertRanges = arrayOf(
+                        ListChangeSet.Range(0, 2),
+                        ListChangeSet.Range(6, 2)
+                    )
+                )
+            }
+
+            // Assert multiple ranges are deleted
+            //
+            // objectListField = [<A, B>, C, D, E, F, <G, H>]
+            realm.writeBlocking {
+                (dataset1 + dataset3).forEach { element ->
+                    delete(
+                        query<RealmListContainer>("stringField = $0", element.stringField)
+                            .first()
+                            .find()!!
+                    )
+                }
+            }
+
+            c.receive().let { resultsChange ->
+                assertIs<UpdatedResults<*>>(resultsChange)
+
+                assertNotNull(resultsChange.list)
+                assertEquals(dataset2.size, resultsChange.list.size)
+
+                assertIsChangeSet(
+                    (resultsChange as UpdatedResults<*>),
+                    deletionRanges = arrayOf(
+                        ListChangeSet.Range(0, 2),
+                        ListChangeSet.Range(6, 2)
+                    )
+                )
+            }
+
+            // Assert a single range is deleted
+            //
+            // objectListField = [<A, B>]
+            realm.writeBlocking {
+                dataset2.forEach { element ->
+                    delete(
+                        query<RealmListContainer>("stringField = $0", element.stringField)
+                            .first()
+                            .find()!!
+                    )
+                }
+            }
+
+            c.receive().let { resultsChange ->
+                assertIs<UpdatedResults<*>>(resultsChange)
+
+                assertNotNull(resultsChange.list)
+                assertTrue(resultsChange.list.isEmpty())
+
+                assertIsChangeSet(
+                    (resultsChange as UpdatedResults<*>),
+                    deletionRanges = arrayOf(
+                        ListChangeSet.Range(0, 4)
+                    )
+                )
+            }
+
+            // Add some values to change
+            //
+            // objectListField = [<C, D, E, F>]
+            realm.writeBlocking {
+                dataset2.forEach {
+                    copyToRealm(it)
+                }
+            }
+
+            c.receive().let { resultsChange ->
+                assertIs<UpdatedResults<*>>(resultsChange)
+
+                assertNotNull(resultsChange.list)
+                assertEquals(dataset2.size, resultsChange.list.size)
+            }
+
+            // Change contents of two ranges of values
+            //
+            // objectListField = [<A>, <B>, E, <F>]
+            realm.writeBlocking {
+                query<RealmListContainer>()
+                    .sort("stringField")
+                    .find { queriedList ->
+                        queriedList[0].stringField = "A"
+                        queriedList[1].stringField = "B"
+                        queriedList[3].stringField = "F"
+                    }
+            }
+
+            c.receive().let { resultsChange ->
+                assertIs<UpdatedResults<*>>(resultsChange)
+
+                assertNotNull(resultsChange.list)
+                assertEquals(dataset2.size, resultsChange.list.size)
+
+                assertIsChangeSet(
+                    (resultsChange as UpdatedResults<*>),
+                    changesRanges = arrayOf(
+                        ListChangeSet.Range(0, 2),
+                        ListChangeSet.Range(3, 1),
+                    )
+                )
+            }
+
             observer.cancel()
             c.close()
         }
@@ -104,36 +271,47 @@ class RealmResultsNotificationsTests : NotificationTests {
     @Test
     override fun cancelAsFlow() {
         runBlocking {
-            val c1 = Channel<RealmResults<Sample>>(1)
-            val c2 = Channel<RealmResults<Sample>>(1)
+            val c1 = Channel<ResultsChange<Sample>>(1)
+            val c2 = Channel<ResultsChange<Sample>>(1)
+
+            realm.write {
+                copyToRealm(Sample().apply { stringField = "Bar" })
+            }
+
             val observer1 = async {
                 realm.query<Sample>()
                     .asFlow()
-                    .filterNot {
-                        it.isEmpty()
-                    }.collect {
+                    .collect {
                         c1.trySend(it)
                     }
             }
             val observer2 = async {
                 realm.query<Sample>()
                     .asFlow()
-                    .filterNot {
-                        it.isEmpty()
-                    }.collect {
+                    .collect {
                         c2.trySend(it)
                     }
             }
-            realm.write {
-                copyToRealm(Sample().apply { stringField = "Bar" })
+
+            c1.receive().let { resultsChange ->
+                assertIs<InitialResults<*>>(resultsChange)
+                assertEquals(1, resultsChange.list.size)
             }
-            assertEquals(1, c1.receive().size)
-            assertEquals(1, c2.receive().size)
+            c2.receive().let { resultsChange ->
+                assertIs<InitialResults<*>>(resultsChange)
+                assertEquals(1, resultsChange.list.size)
+            }
+
             observer1.cancel()
+
             realm.write {
                 copyToRealm(Sample().apply { stringField = "Baz" })
             }
-            assertEquals(2, c2.receive().size)
+
+            c2.receive().let { resultsChange ->
+                assertIs<UpdatedResults<*>>(resultsChange)
+                assertEquals(2, resultsChange.list.size)
+            }
             assertTrue(c1.isEmpty)
             observer2.cancel()
             c1.close()
@@ -144,7 +322,7 @@ class RealmResultsNotificationsTests : NotificationTests {
     @Test
     override fun deleteObservable() {
         runBlocking {
-            val c = Channel<RealmResults<Sample>>(1)
+            val c = Channel<ResultsChange<Sample>>(1)
             realm.write {
                 copyToRealm(
                     Sample().apply {
@@ -159,7 +337,10 @@ class RealmResultsNotificationsTests : NotificationTests {
                         c.trySend(it)
                     }
             }
-            assertEquals(1, c.receive().size)
+            c.receive().let { resultsChange ->
+                assertIs<InitialResults<*>>(resultsChange)
+                assertEquals(1, resultsChange.list.size)
+            }
             realm.write {
                 query<Sample>()
                     .first()
@@ -168,7 +349,10 @@ class RealmResultsNotificationsTests : NotificationTests {
                         delete(sample)
                     }
             }
-            assertEquals(0, c.receive().size)
+            c.receive().let { resultsChange ->
+                assertIs<UpdatedResults<*>>(resultsChange)
+                assertTrue(resultsChange.list.isEmpty())
+            }
             observer.cancel()
             c.close()
         }
@@ -185,7 +369,7 @@ class RealmResultsNotificationsTests : NotificationTests {
                     .asFlow()
                     .collect {
                         when (counter.incrementAndGet()) {
-                            1 -> c.trySend(it.size)
+                            1 -> c.trySend(it.list.size)
                             2 -> {
                                 realm.close()
                                 c.trySend(-1)
@@ -198,7 +382,7 @@ class RealmResultsNotificationsTests : NotificationTests {
                 realm.query<Sample>()
                     .asFlow()
                     .collect {
-                        println(it.first().stringField)
+                        println(it.list.first().stringField)
                         println("$it -> ${realm.isClosed()}")
                     }
             }
@@ -224,9 +408,9 @@ class RealmResultsNotificationsTests : NotificationTests {
                 realm.query<Sample>()
                     .asFlow()
                     .filterNot {
-                        it.isEmpty()
+                        it.list.isEmpty()
                     }.collect {
-                        c.send(it.size)
+                        c.send(it.list.size)
                     }
                 fail("Flow should not be canceled.")
             }